Move DriveTrain interfaces into their own package
