+++ ARCCore/src/main/com/andoverrobotics/core/drivetrain/MecanumDriveTrain.java
@@ -1,4 +1,4 @@
-package com.andoverrobotics.core;
+package com.andoverrobotics.core.drivetrain;
 
 /**
  * Created by Lake Yin on 5/23/2018.
